Level Transistion, Rooms Must Be Cleared to Proceed, and Zombie Spawn Instructions
Added an interactable object for level transistions. If the current room is cleared, the player can press e to proceed to the next level. Players can no longer speed run through rooms without killing any zombies. If a room is not cleared and the current room is not cleared then the player cannot go through the door. This is signified by a black light. Added zombie spawn instructions for rooms which can be loaded by either a zombie spawner placed in a room or an enemy. Zombies now always spawn at the same position when the player enters a room.